refactor(core): extract leaf manifest store module
Hoist the globalThis manifest cache accessors and pure fs and URL
helpers out of registry.ts and class-registration.ts into a new leaf
module at packages/core/src/manifest/store.ts. The new module has no
dependency on ObjectRegistry, registry.ts, or class-registration.ts.
Before this commit, registry.ts kept its own copies of the sync helpers
because it could not import from manifest-loader.ts due to the existing
module cycle. Centralizing the helpers in a new leaf below both ends of
the cycle lets both files import from it without recreating the
circular-init problem. Net change: store.ts adds about 605 lines;
registry.ts sheds about 470 lines of duplicated helpers; the third
duplicate of cloneManifestSchemaColumns goes away; registry.js shrinks
from 163kB to 84kB. No behavior change. 1520 core tests pass.
